Repurposed for release 1.0 added test
